Add rest api delete

+++ src/main/java/ma/youcode/crudSpringB/dao/EmployeeDao.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface EmployeeDao {
     public List<Employee> findAll();
     public void saveEmployee(Employee employee);
-    public Employee readOneEmployee(double idEmployee);
-    public void deleteEmployee(double idEmployee);
+    public Employee readOneEmployee(Long idEmployee);
+    public void deleteEmployee(Long idEmployee);
 
 
 }
